refactor(ui): break compliance.tsx into tab components #264
Split the 1557-line compliance.tsx monolith into focused modules under
src/app/components/compliance/, each under 400 lines. Pure refactor with
zero UI changes — verified by successful production build.
