+++ src/app/components/compliance/vulnerability-modal.tsx
@@ -1,0 +1,270 @@
+import { useState } from "react";
+import { X } from "lucide-react";
+import type { VulnSeverity } from "../../../lib/mock-data/compliance-data";
+import { ModalOverlay } from "./compliance-modals";
+
+// =============================================================================
+// Create Vulnerability Modal
+// =============================================================================
+
+export interface VulnFormData {
+  cveId: string;
+  title: string;
+  description: string;
+  severity: VulnSeverity;
+  cvssScore: number;
+  affectedDevices: number;
+  patchAvailable: boolean;
+}
+
+export function CreateVulnerabilityModal({
+  onClose,
+  onSubmit,
+}: {
+  onClose: () => void;
+  onSubmit: (data: VulnFormData) => void;
+}) {
+  const [cveId, setCveId] = useState("");
+  const [title, setTitle] = useState("");
+  const [description, setDescription] = useState("");
+  const [severity, setSeverity] = useState<VulnSeverity | "">("");
+  const [cvssScore, setCvssScore] = useState("");
+  const [affectedDevices, setAffectedDevices] = useState("");
+  const [patchAvailable, setPatchAvailable] = useState(false);
+  const [errors, setErrors] = useState<Record<string, string>>({});
+
+  const handleSubmit = () => {
+    const newErrors: Record<string, string> = {};
+    if (!cveId.trim()) newErrors.cveId = "CVE ID is required";
+    if (!title.trim()) newErrors.title = "Title is required";
+    if (!severity) newErrors.severity = "Severity is required";
+    if (
+      !cvssScore.trim() ||
+      isNaN(Number(cvssScore)) ||
+      Number(cvssScore) < 0 ||
+      Number(cvssScore) > 10
+    ) {
+      newErrors.cvssScore = "CVSS score must be 0-10";
+    }
+    if (!affectedDevices.trim() || isNaN(Number(affectedDevices)) || Number(affectedDevices) < 0) {
+      newErrors.affectedDevices = "Enter a valid device count";
+    }
+
+    if (Object.keys(newErrors).length > 0) {
+      setErrors(newErrors);
+      return;
+    }
+
+    onSubmit({
+      cveId,
+      title,
+      description,
+      severity: severity as VulnSeverity,
+      cvssScore: Number(cvssScore),
+      affectedDevices: Number(affectedDevices),
+      patchAvailable,
+    });
+  };
+
+  const inputClass =
+    "w-full rounded border border-border bg-card px-2.5 py-1.5 text-sm text-foreground placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-ring/50 focus:border-accent-text";
+
+  return (
+    <ModalOverlay onClose={onClose}>
+      <div className="w-full max-w-md rounded-lg bg-card shadow-xl border border-border">
+        <div className="flex items-center justify-between border-b border-border px-5 py-3">
+          <h2 className="text-[15px] font-semibold text-foreground">Report Vulnerability</h2>
+          <button
+            onClick={onClose}
+            className="rounded p-1 text-muted-foreground hover:text-muted-foreground hover:bg-muted"
+          >
+            <X className="h-4 w-4" />
+          </button>
+        </div>
+
+        <div className="space-y-3 p-5">
+          <div>
+            <label
+              htmlFor="comp-vuln-cve"
+              className="mb-1 block text-[13px] font-semibold text-foreground/80"
+            >
+              CVE ID <span className="text-red-500">*</span>
+            </label>
+            <input
+              id="comp-vuln-cve"
+              type="text"
+              placeholder="CVE-2026-XXXX"
+              value={cveId}
+              onChange={(e) => {
+                setCveId(e.target.value);
+                setErrors((prev) => ({ ...prev, cveId: "" }));
+              }}
+              className={inputClass}
+            />
+            {errors.cveId && (
+              <p className="mt-0.5 text-[12px] text-red-600" role="alert">
+                {errors.cveId}
+              </p>
+            )}
+          </div>
+
+          <div>
+            <label
+              htmlFor="comp-vuln-title"
+              className="mb-1 block text-[13px] font-semibold text-foreground/80"
+            >
+              Title <span className="text-red-500">*</span>
+            </label>
+            <input
+              id="comp-vuln-title"
+              type="text"
+              placeholder="Brief vulnerability title"
+              value={title}
+              onChange={(e) => {
+                setTitle(e.target.value);
+                setErrors((prev) => ({ ...prev, title: "" }));
+              }}
+              className={inputClass}
+            />
+            {errors.title && (
+              <p className="mt-0.5 text-[12px] text-red-600" role="alert">
+                {errors.title}
+              </p>
+            )}
+          </div>
+
+          <div>
+            <label
+              htmlFor="comp-vuln-desc"
+              className="mb-1 block text-[13px] font-semibold text-foreground/80"
+            >
+              Description
+            </label>
+            <textarea
+              id="comp-vuln-desc"
+              placeholder="Detailed description of the vulnerability..."
+              value={description}
+              onChange={(e) => setDescription(e.target.value)}
+              rows={2}
+              className={inputClass}
+            />
+          </div>
+
+          <div className="grid grid-cols-2 gap-3">
+            <div>
+              <label
+                htmlFor="comp-vuln-severity"
+                className="mb-1 block text-[13px] font-semibold text-foreground/80"
+              >
+                Severity <span className="text-red-500">*</span>
+              </label>
+              <select
+                id="comp-vuln-severity"
+                value={severity}
+                onChange={(e) => {
+                  setSeverity(e.target.value as VulnSeverity);
+                  setErrors((prev) => ({ ...prev, severity: "" }));
+                }}
+                className={inputClass}
+              >
+                <option value="">Select severity</option>
+                <option value="Critical">Critical</option>
+                <option value="High">High</option>
+                <option value="Medium">Medium</option>
+                <option value="Low">Low</option>
+                <option value="Info">Info</option>
+              </select>
+              {errors.severity && (
+                <p className="mt-0.5 text-[12px] text-red-600" role="alert">
+                  {errors.severity}
+                </p>
+              )}
+            </div>
+
+            <div>
+              <label
+                htmlFor="comp-vuln-cvss"
+                className="mb-1 block text-[13px] font-semibold text-foreground/80"
+              >
+                CVSS Score <span className="text-red-500">*</span>
+              </label>
+              <input
+                id="comp-vuln-cvss"
+                type="number"
+                step="0.1"
+                min="0"
+                max="10"
+                placeholder="0.0 - 10.0"
+                value={cvssScore}
+                onChange={(e) => {
+                  setCvssScore(e.target.value);
+                  setErrors((prev) => ({ ...prev, cvssScore: "" }));
+                }}
+                className={inputClass}
+              />
+              {errors.cvssScore && (
+                <p className="mt-0.5 text-[12px] text-red-600" role="alert">
+                  {errors.cvssScore}
+                </p>
+              )}
+            </div>
+          </div>
+
+          <div>
+            <label
+              htmlFor="comp-vuln-devices"
+              className="mb-1 block text-[13px] font-semibold text-foreground/80"
+            >
+              Affected Devices <span className="text-red-500">*</span>
+            </label>
+            <input
+              id="comp-vuln-devices"
+              type="number"
+              min="0"
+              placeholder="Number of affected devices"
+              value={affectedDevices}
+              onChange={(e) => {
+                setAffectedDevices(e.target.value);
+                setErrors((prev) => ({ ...prev, affectedDevices: "" }));
+              }}
+              className={inputClass}
+            />
+            {errors.affectedDevices && (
+              <p className="mt-0.5 text-[12px] text-red-600" role="alert">
+                {errors.affectedDevices}
+              </p>
+            )}
+          </div>
+
+          <div className="flex items-center gap-2">
+            <input
+              type="checkbox"
+              id="patch-available"
+              checked={patchAvailable}
+              onChange={(e) => setPatchAvailable(e.target.checked)}
+              className="accent-[#FF7900] h-3.5 w-3.5"
+            />
+            <label htmlFor="patch-available" className="text-[13px] font-medium text-foreground/80">
+              Patch available
+            </label>
+          </div>
+        </div>
+
+        <div className="flex items-center justify-end gap-2 border-t border-border px-5 py-3">
+          <button
+            onClick={onClose}
+            className="rounded border border-border bg-card px-3 py-1.5 text-sm font-medium text-foreground/80 hover:bg-muted transition-colors"
+          >
+            Cancel
+          </button>
+          <button
+            onClick={handleSubmit}
+            className="rounded bg-accent px-3 py-1.5 text-sm font-semibold text-white hover:bg-accent-hover transition-colors"
+          >
+            Create Vulnerability
+          </button>
+        </div>
+      </div>
+    </ModalOverlay>
+  );
+}
