+++ src/app/components/compliance/vulnerability-tab.tsx
@@ -1,0 +1,131 @@
+import { Plus, AlertTriangle } from "lucide-react";
+import { cn } from "../../../lib/utils";
+import type { Role } from "../../../lib/rbac";
+import type { Vulnerability } from "../../../lib/mock-data/compliance-data";
+import { VulnerabilitySearch } from "../search/vulnerability-search";
+import { SEVERITY_CONFIG, REMEDIATION_STYLES, canCreateVulnerability } from "./compliance-shared";
+
+// =============================================================================
+// Vulnerabilities Tab (Card Grid)
+// =============================================================================
+
+interface VulnerabilitiesTabProps {
+  vulnerabilities: Vulnerability[];
+  role: Role;
+  onCreateVuln: () => void;
+}
+
+export function VulnerabilitiesTab({
+  vulnerabilities,
+  role,
+  onCreateVuln,
+}: VulnerabilitiesTabProps) {
+  return (
+    <div className="space-y-3">
+      <div className="flex items-center justify-between">
+        <p className="text-sm text-muted-foreground">
+          {vulnerabilities.length} vulnerabilities sorted by CVSS score (highest first)
+        </p>
+        {canCreateVulnerability(role) && (
+          <button
+            onClick={onCreateVuln}
+            className="flex items-center gap-1.5 rounded bg-accent px-3 py-1.5 text-sm font-semibold text-white hover:bg-accent-hover transition-colors duration-150"
+          >
+            <Plus className="h-3.5 w-3.5" />
+            Report Vulnerability
+          </button>
+        )}
+      </div>
+
+      {/* Story 18.4 — Vulnerability Search with severity filtering */}
+      <VulnerabilitySearch />
+
+      <div className="grid grid-cols-1 gap-3 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
+        {vulnerabilities.map((vuln) => {
+          const sevCfg = SEVERITY_CONFIG[vuln.severity];
+          return (
+            <div
+              key={vuln.id}
+              className={cn(
+                "card-elevated rounded-lg border p-4 space-y-3 transition-all duration-150 hover:shadow-md",
+                vuln.severity === "Critical" && "border-red-200 ring-1 ring-red-100",
+                vuln.severity === "High" && "border-orange-200",
+              )}
+            >
+              {/* Header */}
+              <div className="flex items-start justify-between">
+                <span className="font-mono text-[14px] font-semibold text-foreground">
+                  {vuln.cveId}
+                </span>
+                <span
+                  className={cn(
+                    "rounded-full px-2 py-0.5 text-[12px] font-bold",
+                    sevCfg.bg,
+                    sevCfg.text,
+                  )}
+                >
+                  {vuln.severity}
+                </span>
+              </div>
+
+              {/* Title */}
+              <p className="text-sm text-foreground/80 leading-relaxed line-clamp-2">
+                {vuln.title}
+              </p>
+
+              {/* Stats */}
+              <div className="flex items-center gap-3 text-[12px]">
+                <div className="flex items-center gap-1">
+                  <span className="font-semibold text-foreground">CVSS</span>
+                  <span
+                    className={cn(
+                      "font-bold tabular-nums",
+                      vuln.cvssScore >= 7.0
+                        ? "text-red-600"
+                        : vuln.cvssScore >= 4.0
+                          ? "text-amber-600"
+                          : "text-muted-foreground",
+                    )}
+                  >
+                    {vuln.cvssScore.toFixed(1)}
+                  </span>
+                </div>
+                <div className="h-3 w-px bg-muted" />
+                <div className="flex items-center gap-1">
+                  <AlertTriangle className="h-3 w-3 text-muted-foreground" />
+                  <span className="text-muted-foreground">
+                    {vuln.affectedDevices.toLocaleString()} devices
+                  </span>
+                </div>
+                <div className="h-3 w-px bg-muted" />
+                <span
+                  className={cn(
+                    "font-medium",
+                    vuln.patchAvailable ? "text-emerald-600" : "text-red-600",
+                  )}
+                >
+                  {vuln.patchAvailable ? "Patch available" : "No patch"}
+                </span>
+              </div>
+
+              {/* Remediation */}
+              <div className="flex items-center justify-between pt-1 border-t border-border/60">
+                <span
+                  className={cn(
+                    "rounded-full px-2 py-0.5 text-[12px] font-medium",
+                    REMEDIATION_STYLES[vuln.remediationStatus],
+                  )}
+                >
+                  {vuln.remediationStatus}
+                </span>
+                {vuln.resolvedDate && (
+                  <span className="text-[12px] text-muted-foreground">{vuln.resolvedDate}</span>
+                )}
+              </div>
+            </div>
+          );
+        })}
+      </div>
+    </div>
+  );
+}
